reafactor: move send_batch_inclusion_data_responses to lib
